+++ src/main/java/dev/smootheez/elytracontrol/config/ElytraControlConfig.java
@@ -1,17 +1,11 @@
 package dev.smootheez.elytracontrol.config;
 
 import dev.smootheez.elytracontrol.*;
-import dev.smootheez.elytracontrol.config.option.*;
 import dev.smootheez.scl.api.*;
 import dev.smootheez.scl.config.*;
 
 @Config(name = Constants.MOD_ID, gui = true)
 public class ElytraControlConfig {
-    public static final ConfigOption<Boolean> elytraLock = ConfigOption.create("elytraLock", true);
-    public static final ConfigOption<Boolean> elytraCancel = ConfigOption.create("elytraCancel", true);
-    public static final ConfigOption<Boolean> easyFlight = ConfigOption.create("easyFlight", false);
-    public static final ConfigOption<Boolean> easyFlightMessage = ConfigOption.create("easyFlightMessage", true);
-    public static final ConfigOption<Boolean> elytraLockMessage = ConfigOption.create("elytraLockMessage", true);
-
-    public static final ConfigOption<LockIconMode> lockIconMode = ConfigOption.create("iconLockMode", LockIconMode.ICON_TEXT);
+    public static final ConfigOption<Boolean> DEFAULT_ELTRA_CONTROL = ConfigOption.create("defaultElytraControl", true);
+    public static final ConfigOption<Boolean> ALLOW_FLYING = ConfigOption.create("allowFlying", true);
 }