+++ src/main/java/dev/smootheez/elytracontrol/mixin/LocalPlayerMixin.java
@@ -1,0 +1,8 @@
+package dev.smootheez.elytracontrol.mixin;
+
+import net.minecraft.client.player.*;
+import org.spongepowered.asm.mixin.*;
+
+@Mixin(LocalPlayer.class)
+public abstract class LocalPlayerMixin {
+}