+++ src/main/java/dev/smootheez/elytracontrol/mixin/PlayerEntityMixin.java
@@ -1,8 +1,0 @@
-package dev.smootheez.elytracontrol.mixin;
-
-import net.minecraft.world.entity.player.*;
-import org.spongepowered.asm.mixin.*;
-
-@Mixin(Player.class)
-public abstract class PlayerEntityMixin {
-}